construct: typing: Container
`Container` is a static typing nightmare as it acts as a `dict[str,
Any]`. In addition to that instances are also assigned additional
variables using arbitrary types. Basically `Any` goes in, `Any` comes
out with is not useful for typing.

As we know the types used by `pyelftools`, we can give `mypy` et.al.
some hints on those use cases, but have to put theme here on
`Container` as that is the base for everything.

Yikes.

Signed-off-by: Philipp Hahn <phahn-oss@avm.de>